requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
beautifulsoup4==4.12.3
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
cssselect2==0.7.0
Django==3.2
django-admin-interface==0.18.7
django-admin-list-filter-dropdown==1.0.3
django-admin-rangefilter==0.8.1
django-auditlog==3.0.0
django-bootstrap-v5==1.0.2
django-colorfield==0.6.3
django-cors-headers==3.10.1
django-crispy-forms==1.11.2
django-excel-response==2.0.5
django-filter==2.4.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-polymorphic==3.1.0
django-session-timeout==0.1.0
django-simple-history==3.0.0
django-smart-selects==1.5.9
django-user-agents==0.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
drf-flex-fields==1.0.2
drf-writable-nested==0.7.0
et-xmlfile==1.1.0
freetype-py==2.3.0
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
gunicorn==20.1.0
html2docx==1.6.0
html5lib==1.1
humanize==3.12.0
idna==3.6
jdcal==1.4.1
lxml==5.1.0
num2words==0.5.10
openpyxl==2.4.11
oscrypto==1.3.0
pendulum==3.0.0
pillow==10.2.0
promise==2.3
psycopg2==2.9.9
pycairo==1.25.1
pycparser==2.21
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pypdf==3.17.4
PyPDF2==3.0.1
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-docx==1.1.2
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.9
requests==2.31.0
rlPyCairo==0.3.0
setuptools==69.0.3
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
svglib==1.5.1
text-unidecode==1.3
time-machine==2.13.0
tinycss2==1.2.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
ua-parser==0.18.0
uritools==4.0.2
urllib3==2.1.0
user-agents==2.2.0
webencodings==0.5.1
whitenoise==6.5.0
xhtml2pdf==0.2.13
xlwt==1.3.0
